requirements.txt
Pillow
psutil
piexif
